ahilya-rakshasutra-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
SQLAlchemy>=2.0,<3.0
pydantic>=2.6,<3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
requests==2.32.3
